Update syntax for MacOS
The term for preferences was changed to settings in MacOS Ventura. The
current version of the guides does still references preferences which
might cause confusion to any using the guide.

Let's update the term in the guide from preferences to settings to
prevent confusion when anyone is following the guide.